feat(otterbrix): migrate to otterbrix 1.0.0a13-rc-2 with OID-centric catalog (#20)

+++ frontend/postgres_server/connection/packet_processing.cpp
@@ -466,7 +466,7 @@ namespace frontend::postgres {
         // TODO: PortalSuspended
         int32_t rows_cnt = sdata_result.chunk.size();
         if (limit != 0) {
-            rows_cnt = std::min(static_cast<size_t>(limit), sdata_result.chunk.size());
+            rows_cnt = std::min(static_cast<uint64_t>(limit), sdata_result.chunk.size());
         }
 
         std::vector<std::vector<uint8_t>> response;
